feat(mv3-part-4): implement a passthrough event manager for use with non-background contexts (#5591)
#### Details

This PR changes our browser event handling strategy for non-background contexts to use a new `PassthroughBrowserEventManager` instead of going through the same event management logic that the background requires.

##### Motivation

A followup PR (#5538) involves enabling sharing of some browser event manager code (for creating fnf delays/etc) with the background interpreter in a way that is easiest if it can be allowed to use background-only types. This made it awkward to create the event manager in non-background contexts.

This also simplifies most of the context initializers substantially; it enables moving the interesting logic out of the initializers (which are hard to test) and into separate components (which now have tests in front of them.

Finally, it removes some risk/complexity from places where it isn't needed. For example, doing this earlier would have prevented the issue #5582 fixed.

##### Context

This code was extracted from #5538, since it was getting uncomfortably large.

Technically, we could also use the passthrough event manager from the mv2 background page context. I chose to keep the background event manager there to minimize differences between mv2 and mv3, for the sake of catching issues with the event manager earlier.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [ ] Ran `yarn fastpass`
- [ ] Added/updated relevant unit test(s) (and ran `yarn test`)
- [ ] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [ ] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS