feat(mv3-part-4): improve browser message response type safety (#5626)
#### Details

This PR extends/replaces #5612. It introduces new wrapper types for Browser message handlers/responses (see `browser-message-handler.ts`) which enable catching the type of issue we encountered in #5602 as a type check failure, and it updates all of our Browser message listeners and adapters to use it. It also updates the existing similar response types used in `PostMessageContentHandler` and `Interpreter` to be consistent with the browser message handler types.

Reviewers should give extra attention to the `PostMessageContentHandler` changes, I modified how it worked more substantially/functionally than related code.

Manually smoke tested all pages of the web extension with the changes.

##### Motivation

Prevent the underlying class of issue we fixed one instance of in #5602

##### Context

n/a

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS