fix: Use globalThis in ExceptionTelemetryListener instead of window (#5659)
#### Details

Replace `window` in ExceptionTelemetryListener with `globalThis`.

##### Motivation

Fixes a bug in the MV3 version of the extension. ExceptionTelemetryListener is referenced by SendingExceptionTelemetryListener, which is used in the service worker. Since `window` doesn't exist in the service worker, attempting to initialize SendingExceptionTelemetryListener crashes the service worker initialization script. Using `globalThis` instead will keep the same functionality in non-background contexts while fixing the error in the service worker.