fix: Store last focused element for focus trap detector (#5705)
#### Details

- Extracts TabStopsOrchestrator's keydown listener into its own class, FocusTrapsKeydownHandler
- Store the last focused element in FocusTrapsKeydownHandler and use that instead of dom.activeElement to determine whether we've reached a focus trap

##### Motivation

Fixes a race condition in TabStopsOrchestrator that causes a focus trap false positive. The root cause is that, in the keydown listener, we assume that the focus has not yet changed and use dom.activeElement as the "last focused element." However, sometimes the keyboard listener runs after focus has already changed, in which case the extension wrongly detects that the previous and currently focused elements are the same, and reports a focus trap.

This PR fixes the race condition by storing the last focused element in an instance variable that is updated after every tab press, and using that instead of dom.activeElement.

##### Context

I first tried using TabStopsOrchestrator's existing latestVisitedTabStop variable, but that created a similar race condition: latestVisitedTabStop is changed by the onfocus listener, which may run either before or after the keydown listener. Since each listener must track its previous tab stop separately, I thought it would be clearer to refactor the keydown listener into a new class with its own state instead of adding another variable with a confusingly similar name to TabStopsOrchestrator.