refactor(mv3-part-5): replace ipc AsyncAction with new flux AsyncAction (#5756)
#### Details

Removes electron/ipc/async-action.ts and replace references with common/flux/async-action

##### Motivation

Prior to #5743, there existed an AsyncAction class used in one place in the Electron app. This class served the same function as the new AsyncAction introduced in #5743, so this PR replaces the old class with the new one to avoid naming confusion.