refactor(mv3-part-5): MockInterpreter (#5774)
#### Details

Replace `createInterpreterMock` with a new `MockInterpreter` class in unit tests

##### Motivation

Before this PR, most tests for -ActionCreator classes use `createInterpreterMock`, which calls a callback as it is registered in the interpreter with `registerTypeToPayloadCallback`. This works fine for synchronous callbacks, but there is no way to await an async call back with this strategy, since `registerTypeToPayloadCallback` is synchronous. The few ActionCreators that register async callbacks just use `awaitSettledPromises()` to ensure that the async callback completes.

This PR creates a mock wrapper class that separates callback registration from actually calling the callback, so async callbacks can be awaited without using `awaitSettledPromises`. This also makes the sequence of events clearer in the tests.