feat(mv3-part-5): Convert AssessmentActions to AsyncActions (#5782)
#### Details
Convert all SyncActions in AssessmentActions to AsyncActions

##### Motivation
Feature work

##### Context
Some of the messages that AssessmentActionCreator listens for also have other listeners registered in ActionCreator. Messages are interpreted concurrently using both the global context and the tab context, and since some of the messages affected by this change have listeners in both contexts, some of the now-async actions may be invoked concurrently. This shouldn't be a problem, except that the current Action classes have built-in measures to prevent two actions from running at the same time in the same "scope." To avoid spurious exceptions, this PR adds a "scope" parameter to Action.invoke() calls inside listeners that may run concurrently with other listeners (i.e. listeners for messages that have multiple listeners registered).