fix: let tab stops run if frame messaging fails (#5933)
#### Details
Add a timeout for two method calls that send frame commands for tab stops listener, and continue execution if a timeout is thrown.

##### Motivation
Patch for #5931

##### Context
This is a compromise between the two possible patches mentioned in #5931. We wait 500ms per call in the hopes that the frame message completes in that time, but if it doesn't, we continue execution instead of waiting for the full 30 second frame messaging timeout. Ideally, however, we would be able to tell before messaging the frame whether we expect to get a response.