feat(implement-new-functionality-links): Allow step-specific telemetry (#5980)
#### Details

Our feature adds step-specific telemetry to the `CHANGE_OVERALL_REQUIREMENT_STATUS` data point. This PR adds an optional `Requirement.getCompletedStepDetailsForTelemetry` method that gets queried to create additional telemetry for that `Requirement`. The new data is placed in a `stepDetails` object (the name is a bit generic because a similar approach could be used to add telemetry to _any_ telemetry data point, not just this one). Telemetry is unmodified unless the optional method is provided. This new method gets a copy of the `AssessmentStoreData` so that it has access to the data when creating the additional telemetry.

##### Motivation

Part of current feature work.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for a later PR to handle? -->
A future PR will actually use this data in the label-in-name assessment for links.

<!-- Were there any alternative approaches you considered? What tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue
- [x] Ran `yarn null:autoadd`
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS