fix(mv3-part-7): make getDispatchUpdateEvent synchronous (#6001)
#### Details

Remove await when background sends store update messages.

##### Motivation

This fixes a bug in canary which caused the details view UI to be
unresponsive for ~30 seconds after the tab stops visualizer was turned
on. The user could toggle the visualizer off (with no effect), but could
not change details view tabs.

##### Context

We don't need to wait for a response from this message because it is
sent from, instead of to, the background service worker. If the service
worker goes idle after sending the store update, the message will still
be processed by the context it is sent to.