fix(mv3-part-7): Fix assessment view updater bug (#6012)
#### Details

Fix mv3 bug in which partially automated requirements pages (ex.
Landmarks > landmark roles) blink between results and spinner before
showing results again.

##### Motivation

Fix mv3 bug

##### Context

The root cause of this bug was that we were sending too many enable
visualization messages. We send these messages in
`assessment-view-update-handler` based on if the assessment store data
thinks that a step has been scanned or not
(`props.assessmentData.testStepStatus[step].isStepScanned`). We were
seeing this bug because:

- The `assessment-view-update-handler` is called every time any store
data updates
- Multiple stores have listeners for the scanCompleted action
- If any store besides the assessment store completes its updates first,
it causes the `assessment-view-update-handler` to run its check against
assessment store data that is in the process of being updates/ already
out of date.
- This can create a situation where the assessment store is finished
with the scan but the `assessment-view-update-handler` thinks it hasn't,
so it re-sends the enable visualization message, which causes a re-scan.

The solution I implemented is to only send enable visualization messages
is the store update came from the assessment store. This means that we
won't be confused by unrelated store updates anymore.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ n/a] Addresses an existing issue: #0000
- [x] Ran `yarn null:autoadd`
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS