fix: improve automated tab stops instance generation by batching getUniqueSelector (#6015)
#### Details

Creates a new axe-util API that allows us to getSelectors based on an
array, meaning setup and teardown of axe is only done once.

##### Motivation

Significantly improves perf when many instances are generated by
automated tab stops.

##### Context

Before:

![canaryTabStops](https://user-images.githubusercontent.com/32555133/189216341-31859de0-ea6b-4a04-ad0c-943a8d7c96a0.gif)

After:

![fixedDevTabStops](https://user-images.githubusercontent.com/32555133/189216375-538c7f3f-4b38-4474-a972-2b575c5cbd3c.gif)

I believed this perf issue existed because, upon stopping, we loop
through each tabbable item we didn't see, in this case hundreds of
tabbable elements, and send a separate action message to create an
instance for each one. Obviously, that wouldn't be performant and we
could batch this. This was kind of correct and does improve performance
but for scenarios which could take 20-30s, it would only improve that by
1-2s.
 
In actuality, the main performance culprit is `getUniqueSelector`.
Basically, we need to setup axe each time we use that API but the cost
of doing so is high. Doing so hundreds of times, like we can do in this
case, is what causes the hanging.

I did not investigate whether there are other places where we use
getUniqueSelector in quick succession as there might be perf gains to be
had.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn null:autoadd`
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [x] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS