fix: register message listener after all stores are initialized (#6040)
#### Details

Move creation of BackgroundMessageListener to the end of
background-init.ts and service-worker-init.ts

##### Motivation

We've been seeing intermittent store errors that cause the extension to
freeze until the target page is refreshed. I believe this is because
we're creating and registering the background message listener before
everything else is fully initialized, so it's possible for the
background to receive a message for a store that has not yet been
initialized.

This PR moves that listener registration later in the script code, after
all stores have been created and initialized. The BrowserEventManager
defers events until a listener is registered, so any messages received
before BackgroundMessageListener is registered will still be processed
after everything is initialized.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn null:autoadd`
- [x] Ran `yarn fastpass`
- [n/a] Added/updated relevant unit test(s) (and ran `yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage
report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS