fix: make DevToolsMonitor more failure-tolerant (#6080)
#### Details

Only mark a devtool as "closed" if we fail to reach it three times in a
row.

##### Motivation

Currently, in the background or service worker, each tab has a
DevToolsMonitor that polls any open devtools until they fail to respond,
and then marks them as closed in the DevToolStore. This means that, if
the browser is running slowly and the devtool misses just one ping, the
background will wrongly assume that devtool has closed. From a user
perspective, this puts the extension in a state where clicking "inspect
element" on the Automated Checks visualizer shows a message for the user
to open a devtool, even though one is already open.

This PR makes devtool monitoring more failure-tolerant by requiring a
devtool to fail three pings in a row before we mark it as closed.