fix: only call createObjectURL when needed (#6095)
#### Details

Currently, the export dialog calls `window.URL.createObjectURL()` on
every rerender, even when the dialog is not visible. This PR moves that
call so it happens right after reports are generated, when the download
button is clicked.

##### Motivation

I've been seeing some performance issues from `createObjectURL`
occasionally taking 1-2 seconds to complete and delaying the UI from
updating. It doesn't happen often, and only in the Manifest V3 build of
the extension. In particular, if I run automated checks on a page with a
lot of iframes (such as cnn.com), and select "start over" immediately
after the scan completes, the UI doesn't update for about a second
because `createObjectURL` runs extremely slowly _only_ in the update
cycle after the visualizer is re-enabled. The lag is less obvious on
faster machines, but it still appears slightly slower than the MV2
extension. This PR fixes that issue.

##### Context

I suspect the lag is the result of some implementation details with
chrome's `window` API, but I don't have enough information to guess what
it could be or why it only happens in Manifest V3.