fix: send automated tab stops results in batches (#6114)
#### Details

Optimizations for tab stops results reporting
- Send automated tab stops results to background in batches instead of
one at a time (minimizes disk writes for store updates)
- Send results between frames in batches instead of one at a time
(minimizes frame messages and calls to getUniqueSelector, which is a
time-consuming synchronous operation)

##### Motivation

Processing automated tab stops results can take a long time, especially
on pages with a lot of iframes and tabbable elements. This can be seen
by opening a page with a lot of iframes and tabbed elements (such as
cnn.com), enabling the tab stops ad hoc visualizer and tabbing through a
few elements, and then disabling it and immediately trying to enable
another visualizer. On my surface laptop, on cnn.com, there is a delay
of 6-8 seconds between clicking the toggle and popup UI updating to show
the spinner, plus another 2-3 seconds before the visualizer shows in the
page.

This PR optimizes result processing by batching the most time-consuming
processes and minimizing the number of messages sent between frames and
between the content script and background. This reduces the 6-8 second
UI delay to less than a second, so enabling a new visualizer after
disabling tab stops takes about the same amount of time as without first
enabling and disabling tab stops.

##### Context

Continuation of tab stops performance improvements from #6015

Co-authored-by: Wahaj Abid <waabid@microsoft.com>