feat(medium-pass-1): add hidden feature flag for MediumPass (#6153)
#### Details

Adds a hidden feature flag for the MediumPass feature.

##### Motivation

Feature work 🚀 

##### Context

This feature creates the feature flag to be used in later features. It
will be used in the launch pad and in the DetailsView Pivot to either
display or hide the Medium Pass option.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn null:autoadd`
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS

Co-authored-by: madalynrose <madparker@microsoft.com>