feat(medium-pass-1): Creates messages, telemetry events, and action creator for Medium Pass (#6152)
#### Details

Creates messages, telemetry events and the action creator for Medium
Pass. They are largely duplicates of the Assessment versions barring
minor changes:
- some messages referenced details view, that is now gone.
- removed an action message that was for detailsViewInitialized; this
will be refactored in a future PR from AssessmentActionCreator to the
general ActionCreator.

##### Motivation

Feature work

##### Context

No wiring up is done in this PR. No renaming of Assessment to MediumPass
is done for the action creator methods as the actions cannot be renamed
and so, I have left it as is, for now, which should hopefully make a
review easier as it can be easily compared with the Assessment Action
Creator.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn null:autoadd`
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS