refactor(medium-pass-1): refactor enableTest to not use getStoreData (#6169)
#### Details

Use `ScanData` instead of calling getStoreData in enableTest.

##### Motivation

Feature work.

##### Context

This limits how many properties we may have to overwrite for MediumPass;
without this refactor, `enableTest` would also need to be overwritten as
enableTest references the `getStoreData` function from Assessment
Builder (and not the function on the config object itself). The typings
also now match the other methods like `disableTest` and `getTestStatus`.

Note: visualization-store.test.ts is not modified because ... it relies
on using the actual visualization configuration factory as opposed to a
mock. Modifying that would require significantly more work outside of
the scope of this PR but did verify that visualization store is doing
what we're expected. That is to say, the unit test verifies the state is
being changed, it does not care how it is done (which is usually fine
... but there's underlying dependency on the actual config factory which
isn't great).

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn null:autoadd`
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS