refactor(medium-pass-1): add forEachConfig and medium pass provider to VisualizationConfigurationFactory (#6202)
#### Details

Heavily suggest reviewing by commit due to signature changes causing
many files to be changed.

This PR can roughly be broken into 3 things:

1. Use constructor to pass in assessment and medium pass provider: this
requires the most amount of file changes as each usage/creation of the
factory needs to be changed.
2. Add the ability to loop through each configuration from the factory
instead of components being responsible for this. The benefit here is
that each component doing this no longer needs to be aware of the
difference between assessment and not but also the structure of the
assessment data.
3. Consumption of this forEachConfig method in the appropriate
locations.

##### Motivation

feature work.

##### Context

1. `getIdentifier` is refactored out of AssessmentBuilder. This allows
us to use test mode as a part of generating the identifier allowing for
it to be unique across medium pass and assessment; removing it is the
3rd commit but adding it to config factory is in the first commit.
2. The mediumPass provider is currently optional so as not to cause
issues/errors as the work on that is in the next PR. This will be
removed in that PR as well.
3. I did verify that the async behavior in
`VisualizationStateChangeHandler` was kept the same.
4. There is obviously a lot of potential refactoring work around the
configuration factory (many tests don't mock it, we could build configs
only once, we could break the building from the using/getting, etc.) but
this PR is large enough as it is and I tried to maintain what we were
doing in the main branch as best as I could.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn null:autoadd`
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS